requirements.txt
attrs>=18.2.0
toolz>=0.9.0